requirements.txt
streamlit==1.24.0
streamlit-webrtc
sounddevice
numpy
pandas
azure-cognitiveservices-speech
soundfile
PyAudio